Integrate TachyWooting into TachyPy with native analog feedback support
* Add first-class TachyWooting integration in the TachyPy package.
* Expose `WOOTING_ACQUISITION` through TachyPy for a simpler user-facing API.
* Enable natural analog keyboard workflows directly from TachyPy.
* Add visual pressure feedback for light-press and readiness phases.
* Refactor the pressure feedback system and remove the unused `PressureSource` interface.
* Improve `PressureFeedbackWidget` with better pressure text, geometry, and interactive fixation cross handling.
* Integrate Wooting acquisition with the TachyPy visual feedback system.
* Keep Wooting support as an optional dependency so TachyPy core remains lightweight.
* Update documentation, demos, and tests for the new integration.
* Add tests for pressure feedback rendering and pressure state management.
