+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = tachypy
 version = 0.1.17
-author = Ian Charest and Frederic Gosselin
+author = Ian Charest, Mathias Salvas-Hebert and Frederic Gosselin
 author_email = charest.ian@gmail.com
 description = A package for psychophysics in Python, using GLFW and OpenGL.
 long_description = file: README.md
@@ -41,18 +41,14 @@ test =
     pytest>=7.0
     pytest-cov>=5.0
     ruff>=0.6
-glfw =
-    glfw>=2.7
-audio =
-text =
-    Pillow>=10.0
-system_text =
-    freetype-py>=2.4
-    uharfbuzz>=0.39
+wooting =
+    tachywooting>=0.2.0
 
 [options.entry_points]
 console_scripts =
     tachypy-clock-demo = tachypy.examples.clock_timer_demo:main
+    tachypy-wooting-fixation-demo = tachypy.wooting.demos.visual_fixation_demo:main
+    tachypy-wooting-mini-bw = tachypy.wooting.demos.mini_bw_experiment:main
 
 [tool:pytest]
 addopts = -q
