+++ setup.py
@@ -14,6 +14,7 @@ setup(
     package_dir={'': 'src'},
     install_requires=requires,
     extras_require={
+        'test': ['pytest>=7.0', 'pytest-cov>=5.0', 'ruff>=0.6'],
         'wooting': ['tachywooting>=0.2.0'],
     },
     python_requires='>=3.10',
