+++ src/tachypy/feedback/__init__.py
@@ -1,19 +1,24 @@
 """Visual pressure-feedback toolkit for analog keyboards.
 
-This subpackage is keyboard-agnostic: it renders feedback for any object that
-satisfies :class:`PressureSource` (it never imports a keyboard package). Most
-users do not import from here directly — they call ``wait_light_press_visual``
-on an acquisition class enriched with :class:`VisualPressureFeedbackMixin`
+Keyboard-agnostic: it renders feedback for any object satisfying
+:class:`PressureSource` and never imports a keyboard package. Most users do not
+import from here directly — they call ``wait_light_press_visual`` on an
+acquisition class enriched with :class:`VisualPressureFeedbackMixin`
 (see :mod:`tachypy.wooting`). The building blocks below are exposed for power
 users who want custom widgets or to drive the loop manually.
+
+Layout
+------
+- ``model``   — pure logic: ``PressureSource``, ``PressureFeedbackConfig``
+  (thresholds, hold, and scaling), and ``PressureFeedbackState`` (no OpenGL).
+- ``widgets`` — rendering: ``PressureFeedbackWidget`` (ABC) and the default
+  ``InteractiveFixationCross``.
+- ``runner``  — the agnostic loop (``run_light_press_visual``) and the
+  user-facing ``VisualPressureFeedbackMixin``.
 """
-from .engine import DEFAULT_EXIT_KEYS, run_light_press_visual
-from .fixation import InteractiveFixationCross
-from .mapping import PressureScaleMapper
-from .mixin import VisualPressureFeedbackMixin
-from .source import PressureSource
-from .state import PressureFeedbackConfig, PressureFeedbackState, PressureStatus
-from .widgets import PressureFeedbackWidget
+from .model import PressureFeedbackConfig, PressureFeedbackState, PressureSource
+from .runner import DEFAULT_EXIT_KEYS, VisualPressureFeedbackMixin, run_light_press_visual
+from .widgets import InteractiveFixationCross, PressureFeedbackWidget
 
 __all__ = [
     "DEFAULT_EXIT_KEYS",
@@ -21,9 +26,7 @@ __all__ = [
     "PressureFeedbackConfig",
     "PressureFeedbackState",
     "PressureFeedbackWidget",
-    "PressureScaleMapper",
     "PressureSource",
-    "PressureStatus",
     "VisualPressureFeedbackMixin",
     "run_light_press_visual",
 ]
